Refactoring file logic
